Fix event priority.
`Hal`'s renderer listener is attached with priority 100 too, so the 'ZF\ApiProblem\ApiProblemStrategy' listener is executed before the rendering just because right now the module ApiProblem loade first than the module Hal, and that is not cool. I pick 400 as priority value becuase is important that this listener be executed after the `ApiProblemListener::onRender` listener, and the priority of that listener is 1000.