use JSON_PARTIAL_OUTPUT_ON_ERROR for json_decode()
ApiProblem was not able to render exceptions when the option
ApiProblem::detailIncludesStackTrace was true and the exception contained
PHP streams in the arguments of the traces ("args" key in
Exception::getTraces()). For example: GuzzleHttp works internally with
streams which meant that some exceptions raised from GuzzleHttp didn't
result in a ApiProblem response but in an empty 500 response because
json_encode() failed.